+++ src/main/kotlin/com/chsdngm/tilly/repository/PrivateModeratorRepository.kt
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Repository
 import org.springframework.transaction.annotation.Transactional
 
 @Repository
-interface PrivateModeratorRepository : CrudRepository<PrivateModerator, Int> {
+interface PrivateModeratorRepository : CrudRepository<PrivateModerator, Long> {
   @Query(value = """
     select user_id
     from private_moderator
     where assigned > (NOW() - interval '1 DAY');
   """, nativeQuery = true)
-  fun findCurrentModeratorsIds(): List<Int>
+  fun findCurrentModeratorsIds(): List<Long>
 
   @Transactional
   @Modifying
@@ -23,5 +23,5 @@ interface PrivateModeratorRepository : CrudRepository<PrivateModerator, Int> {
     values (:id, now())
     on conflict on constraint private_moderator_pkey do update set assigned = now()
   """, nativeQuery = true)
-  fun addPrivateModerator(id: Int)
+  fun addPrivateModerator(id: Long)
 }
